requirements.txt
flask>=2.0.1
requests>=2.28.1
web3>=6.0.0
pandas>=1.3.3
python-dotenv>=0.19.0
dune-client>=1.3.0
flask-caching>=2.0.2
gunicorn>=20.1.0